Split out service (servant) api and core api
