+++ app/AugsLink/Service/Application.hs
@@ -8,7 +8,8 @@ import AugsLink.Service.Handlers.GetHome ( home )
 import AugsLink.Service.Handlers.PostHome ( create )
 import AugsLink.Service.Handlers.RoomWs ( enter )
 import AugsLink.Service.Handlers.GetRoom ( room )
-import AugsLink.Service.Room (Registry)
+
+import AugsLink.Core.API
 
 handlers :: CLArgs -> Registry IO -> Server API
 handlers opts rr =
