+++ app/Main.hs
@@ -2,9 +2,9 @@ module Main( main ) where
 
 import Network.Wai.Handler.Warp ( run )
 
+import AugsLink.Core.Room
 import AugsLink.Service.Application ( server )
 import CommandLine ( getCLArgs )
-import AugsLink.Service.Room (newRegistry)
 
 main :: IO ()
 main = do
